Deepak -- intravital -- stable -- InvivoCytometer, DNADamageAnalysis
>> Modified all GUI tools of Invivo Cytometer and DNADamageAnalyzer.m to accept more file formats

For this i write a general class called BioImageReader.m and a function called uiGetBioImageData.m

Experimenting with size features in ComputeRegionMergingFeatures.m

>> DNA Damage Analysis

DNADamageAnalyzer.m now has been extended to 

-- uses a machine learning model to detect foci -- trainFociDetectionModel.m, ComputeFociDetectionFeatures.m
-- segment the drug and macrophage channels and perform some colocalization analysis - segmentDrugCisplatin.m, segmentMacrophages.m

>> Added a new function for thresholding by removing the background using morphological opening - thresholdSBR.m

This works better on varied datasets than the local minimum error thresholding i was using for nuclei segmentation. So i plan to make this the default thresholding algorithm in the future.

>> Updated trainCellStateClassificationModel_3D.m to generate G1SG2_M model as well

>> Added some utility functions -- imresizetotarget.m


git-svn-id: https://orchestra.med.harvard.edu/svn/lccb/matlab/projects/intravital/main@11522 0bcb941b-74d1-0310-bb26-e6636707f451

+++ segmentMacrophages.m
@@ -11,7 +11,7 @@ function [imMacrophageSeg] = segmentMacrophages(imMacrophage, varargin)
     
     PARAMETERS = p.Results;
     
-    imMacrophageAdjusted = matitk('FMEDIAN', round(2 * min(PARAMETERS.spacing) ./ PARAMETERS.spacing), imMacrophage);
+    imMacrophageAdjusted = matitk('FMEDIAN', round(2 * min(PARAMETERS.spacing) ./ PARAMETERS.spacing), double(imMacrophage));
     
     imMacrophageSeg = thresholdSBR(imMacrophageAdjusted, ...
                                    PARAMETERS.maxObjectRadius, PARAMETERS.minSignalToBackgroundRatio, ...
